refactor(agent-cli): move response reshaping into agent_call/reshape.rs under the file-size cap
The merge with main left three modules over R8's 1400-LOC hard cap. Each is
split along the seam its own doc already names, never by trimming comments:

- agent_call.rs 1681 -> 1334, with the agent-layer payload reshaping in
  agent_call/reshape.rs: the paged-list and base64 consts, take_list_page_args,
  paginate_list_reply, base64_byte_fields, reshape_reply (whose doc numbers the
  outbound order) and the inbound fence-strip mirror. The fencing tables stay in
  the parent, which both directions read.
- mcp.rs 1401 -> 1209, with tools/list's catalogue (titles, descriptions,
  annotations, input schemas) in mcp/schemas.rs. The tool-name consts stay:
  tool_argv and classify_tool_call match on them, so they are protocol.
- agent_read.rs 1405 -> 746, by moving its inline test module to
  agent_read/tests.rs. R8 counts a whole file, and this was the last module in
  the tree still carrying tests inline; found_jobs::tests keeps resolving its
  fixtures through the same module path.

No behaviour change. Visibility widens only where a cross-module reader needs
it, mirroring agent_read::found_jobs; the docs and ADR-038 pointers that named
a moved symbol's old file are updated with it.
